Recovered from Eclipse.
git-svn-id: https://subversion.assembla.com/svn/cmosher@1106 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ bindiff/src/BinDiff.java
@@ -87,12 +87,12 @@ public class BinDiff
         statechange(END,0);
     }
 
-	protected void statechange(int newstate, int c)
+	protected void statechange(int newstate, long c)
 	{
 		statechange(newstate,c,null);
 	}
 
-    protected void statechange(int newstate, int c, TellStream f)
+    protected void statechange(int newstate, long c, TellStream f)
     {
         if (newstate == START)
         {
